ai-agent/requirements.txt
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-pinecone>=0.0.3
pinecone-client>=3.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0
pydantic>=2.0.0